[INFOI-1888181][planner] Fix wrong acc type for LongDeclarativeLag1AggFunction

+++ flink-table/flink-table-planner-blink/src/main/java/org/apache/flink/table/planner/functions/aggfunctions/DeclarativeLag1AggFunction.java
@@ -119,7 +119,7 @@ public abstract class DeclarativeLag1AggFunction extends DeclarativeAggregateFun
 	public static class LongDeclarativeLag1AggFunction extends DeclarativeLag1AggFunction {
 		@Override
 		public DataType getResultType() {
-			return DataTypes.INT();
+			return DataTypes.BIGINT();
 		}
 	}
